Rename new plan test directory to tests
